Rename and move the date time handling facility
