+++ src/main/java/tasklist/Deadline.java
@@ -1,15 +1,15 @@
 package tasklist;
 
 import exception.MalformedUserInputException;
-import parser.DateTimeParser;
+import common.DateTimeHandler;
 
 public class Deadline extends Task {
 
-    private DateTimeParser by;
+    private DateTimeHandler by;
 
     public Deadline(String description, String by, boolean isDone) throws MalformedUserInputException {
         super(description, isDone);
-        this.by = new DateTimeParser(by);
+        this.by = new DateTimeHandler(by);
     }
 
     @Override
